+++ rumqttc/src/v5/eventloop.rs
@@ -130,7 +130,15 @@ impl EventLoop {
         self.pending.extend(self.state.clean());
 
         // drain requests from channel which weren't yet received
-        let requests_in_channel = self.requests_rx.drain();
+        let mut requests_in_channel: Vec<_> = self.requests_rx.drain().collect();
+
+        requests_in_channel.retain(|request| {
+            match request {
+                Request::PubAck(_) => false, // Wait for publish retransmission, else the broker could be confused by an unexpected ack
+                _ => true,
+            }
+        });
+
         self.pending.extend(requests_in_channel);
     }
 
@@ -145,18 +153,25 @@ impl EventLoop {
                 connect(&mut self.options, &mut self.state),
             )
             .await??;
+            // Last session might contain packets which aren't acked. If it's a new session, clear the pending packets.
+            if !connack.session_present {
+                self.pending.clear();
+            }
             self.network = Some(network);
 
             if self.keepalive_timeout.is_none() {
                 self.keepalive_timeout = Some(Box::pin(time::sleep(self.options.keep_alive)));
             }
 
-            self.state.handle_incoming_packet(connack)?;
+            self.state
+                .handle_incoming_packet(Incoming::ConnAck(connack))?;
         }
 
         match self.select().await {
             Ok(v) => Ok(v),
             Err(e) => {
+                // MQTT requires that packets pending acknowledgement should be republished on session resume.
+                // Move pending messages from state to eventloop.
                 self.clean();
                 Err(e)
             }
@@ -270,19 +285,14 @@ impl EventLoop {
 async fn connect(
     options: &mut MqttOptions,
     state: &mut MqttState,
-) -> Result<(Network, Incoming), ConnectionError> {
+) -> Result<(Network, ConnAck), ConnectionError> {
     // connect to the broker
     let mut network = network_connect(options).await?;
 
     // make MQTT connection request (which internally awaits for ack)
-    let packet = mqtt_connect(options, &mut network, state).await?;
+    let connack = mqtt_connect(options, &mut network, state).await?;
 
-    // Last session might contain packets which aren't acked. MQTT says these packets should be
-    // republished in the next session
-    // move pending messages from state to eventloop
-    // let pending = self.state.clean();
-    // self.pending = pending.into_iter();
-    Ok((network, packet))
+    Ok((network, connack))
 }
 
 async fn network_connect(options: &MqttOptions) -> Result<Network, ConnectionError> {
@@ -395,7 +405,7 @@ async fn mqtt_connect(
     options: &mut MqttOptions,
     network: &mut Network,
     state: &mut MqttState,
-) -> Result<Incoming, ConnectionError> {
+) -> Result<ConnAck, ConnectionError> {
     let keep_alive = options.keep_alive().as_secs() as u16;
     let clean_start = options.clean_start();
     let client_id = options.client_id();
@@ -414,14 +424,18 @@ async fn mqtt_connect(
     loop {
         match network.read().await? {
             Incoming::ConnAck(connack) if connack.code == ConnectReturnCode::Success => {
-                // Override local keep_alive value if set by server.
                 if let Some(props) = &connack.properties {
                     if let Some(keep_alive) = props.server_keep_alive {
                         options.keep_alive = Duration::from_secs(keep_alive as u64);
                     }
                     network.set_max_outgoing_size(props.max_packet_size);
+
+                    // Override local session_expiry_interval value if set by server.
+                    if props.session_expiry_interval.is_some() {
+                        options.set_session_expiry_interval(props.session_expiry_interval);
+                    }
                 }
-                return Ok(Packet::ConnAck(connack));
+                return Ok(connack);
             }
             Incoming::ConnAck(connack) => {
                 return Err(ConnectionError::ConnectionRefused(connack.code))