+++ rumqttc/src/v5/mqttbytes/v5/auth.rs
@@ -1,23 +1,54 @@
-use super::*;
 use bytes::{Buf, BufMut, Bytes, BytesMut};
 
-/// Auth packet
-#[derive(Clone, Debug, PartialEq, Eq)]
+use super::{
+    len_len, length, property, read_mqtt_bytes, read_mqtt_string, read_u8, write_mqtt_bytes,
+    write_mqtt_string, write_remaining_length, Error, FixedHeader, PropertyType,
+};
+
+/// Auth packet reason code
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum AuthReasonCode {
+    Success,
+    Continue,
+    ReAuthenticate,
+}
+
+impl AuthReasonCode {
+    fn read(bytes: &mut Bytes) -> Result<Self, Error> {
+        let reason_code = read_u8(bytes)?;
+        let code = match reason_code {
+            0x00 => AuthReasonCode::Success,
+            0x18 => AuthReasonCode::Continue,
+            0x19 => AuthReasonCode::ReAuthenticate,
+            _ => return Err(Error::MalformedPacket),
+        };
+
+        Ok(code)
+    }
+
+    fn write(&self, buffer: &mut BytesMut) -> Result<(), Error> {
+        let reason_code = match self {
+            AuthReasonCode::Success => 0x00,
+            AuthReasonCode::Continue => 0x18,
+            AuthReasonCode::ReAuthenticate => 0x19,
+        };
+
+        buffer.put_u8(reason_code);
+
+        Ok(())
+    }
+}
+
+/// Used to perform extended authentication exchange
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Auth {
-    pub reason: AuthReasonCode,
+    pub code: AuthReasonCode,
     pub properties: Option<AuthProperties>,
 }
 
 impl Auth {
-    pub fn new(reason: AuthReasonCode, properties: Option<AuthProperties>) -> Self {
-        Self { reason, properties }
-    }
-
-    pub fn size(&self) -> usize {
-        let len = self.len();
-        let remaining_len_size = len_len(len);
-
-        1 + remaining_len_size + len
+    pub fn new(code: AuthReasonCode, properties: Option<AuthProperties>) -> Self {
+        Self { code, properties }
     }
 
     fn len(&self) -> usize {
@@ -35,24 +66,31 @@ impl Auth {
         len
     }
 
-    pub fn read(fixed_header: FixedHeader, mut bytes: Bytes) -> Result<Auth, Error> {
+    pub fn size(&self) -> usize {
+        let len = self.len();
+        let remaining_len_size = len_len(len);
+
+        1 + remaining_len_size + len
+    }
+
+    pub fn read(fixed_header: FixedHeader, mut bytes: Bytes) -> Result<Self, Error> {
         let variable_header_index = fixed_header.fixed_header_len;
         bytes.advance(variable_header_index);
-        let code = read_u8(&mut bytes)?;
-        let reason = reason(code)?;
+
+        let code = AuthReasonCode::read(&mut bytes)?;
         let properties = AuthProperties::read(&mut bytes)?;
-        let auth = Auth { reason, properties };
+        let auth = Auth { code, properties };
 
         Ok(auth)
     }
 
     pub fn write(&self, buffer: &mut BytesMut) -> Result<usize, Error> {
-        let len = Self::len(self);
         buffer.put_u8(0xF0);
 
+        let len = self.len();
         let count = write_remaining_length(buffer, len)?;
-        buffer.put_u8(code(self.reason));
 
+        self.code.write(buffer)?;
         if let Some(p) = &self.properties {
             p.write(buffer)?;
         } else {
@@ -63,23 +101,11 @@ impl Auth {
     }
 }
 
-/// Return code in auth
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AuthReasonCode {
-    Success,
-    ContinueAuthentication,
-    Reauthenticate,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
+#[derive(Debug, Default, Clone, PartialEq, Eq)]
 pub struct AuthProperties {
-    /// Method of authentication
-    pub authentication_method: Option<String>,
-    /// Authentication data
-    pub authentication_data: Option<Bytes>,
-    /// Reason for disconnection
-    pub reason_string: Option<String>,
-    /// List of user properties
+    pub method: Option<String>,
+    pub data: Option<Bytes>,
+    pub reason: Option<String>,
     pub user_properties: Vec<(String, String)>,
 }
 
@@ -87,36 +113,37 @@ impl AuthProperties {
     fn len(&self) -> usize {
         let mut len = 0;
 
-        if let Some(authentication_method) = &self.authentication_method {
-            len += 1 + 2 + authentication_method.len();
+        if let Some(method) = &self.method {
+            let m_len = method.len();
+            len += 1 + 2 + m_len;
         }
 
-        if let Some(authentication_data) = &self.authentication_data {
-            len += 1 + 2 + authentication_data.len();
+        if let Some(data) = &self.data {
+            let d_len = data.len();
+            len += 1 + 2 + d_len;
         }
 
-        if let Some(reason) = &self.reason_string {
-            len += 1 + 2 + reason.len();
+        if let Some(reason) = &self.reason {
+            let r_len = reason.len();
+            len += 1 + 2 + r_len;
         }
 
         for (key, value) in self.user_properties.iter() {
-            len += 1 + 2 + key.len() + 2 + value.len();
+            let p_len = key.len() + value.len();
+            len += 1 + 4 + p_len;
         }
 
         len
     }
 
     pub fn read(bytes: &mut Bytes) -> Result<Option<AuthProperties>, Error> {
-        let mut authentication_method = None;
-        let mut authentication_data = None;
-        let mut reason_string = None;
-        let mut user_properties = Vec::new();
-
         let (properties_len_len, properties_len) = length(bytes.iter())?;
         bytes.advance(properties_len_len);
         if properties_len == 0 {
             return Ok(None);
         }
+
+        let mut props = AuthProperties::default();
 
         let mut cursor = 0;
         // read until cursor reaches property length. properties_len = 0 will skip this loop
@@ -128,51 +155,46 @@ impl AuthProperties {
                 PropertyType::AuthenticationMethod => {
                     let method = read_mqtt_string(bytes)?;
                     cursor += 2 + method.len();
-                    authentication_method = Some(method);
+                    props.method = Some(method);
                 }
                 PropertyType::AuthenticationData => {
                     let data = read_mqtt_bytes(bytes)?;
                     cursor += 2 + data.len();
-                    authentication_data = Some(data);
+                    props.data = Some(data);
                 }
                 PropertyType::ReasonString => {
                     let reason = read_mqtt_string(bytes)?;
                     cursor += 2 + reason.len();
-                    reason_string = Some(reason);
+                    props.reason = Some(reason);
                 }
                 PropertyType::UserProperty => {
                     let key = read_mqtt_string(bytes)?;
                     let value = read_mqtt_string(bytes)?;
                     cursor += 2 + key.len() + 2 + value.len();
-                    user_properties.push((key, value));
+                    props.user_properties.push((key, value));
                 }
                 _ => return Err(Error::InvalidPropertyType(prop)),
             }
         }
 
-        Ok(Some(AuthProperties {
-            authentication_method,
-            authentication_data,
-            reason_string,
-            user_properties,
-        }))
+        Ok(Some(props))
     }
 
     pub fn write(&self, buffer: &mut BytesMut) -> Result<(), Error> {
         let len = self.len();
         write_remaining_length(buffer, len)?;
 
-        if let Some(authentication_method) = &self.authentication_method {
+        if let Some(authentication_method) = &self.method {
             buffer.put_u8(PropertyType::AuthenticationMethod as u8);
             write_mqtt_string(buffer, authentication_method);
         }
 
-        if let Some(authentication_data) = &self.authentication_data {
+        if let Some(authentication_data) = &self.data {
             buffer.put_u8(PropertyType::AuthenticationData as u8);
             write_mqtt_bytes(buffer, authentication_data);
         }
 
-        if let Some(reason) = &self.reason_string {
+        if let Some(reason) = &self.reason {
             buffer.put_u8(PropertyType::ReasonString as u8);
             write_mqtt_string(buffer, reason);
         }
@@ -184,25 +206,6 @@ impl AuthProperties {
         }
 
         Ok(())
-    }
-}
-
-fn reason(num: u8) -> Result<AuthReasonCode, Error> {
-    let code = match num {
-        0x00 => AuthReasonCode::Success,
-        0x18 => AuthReasonCode::ContinueAuthentication,
-        0x19 => AuthReasonCode::Reauthenticate,
-        num => return Err(Error::InvalidReason(num)),
-    };
-
-    Ok(code)
-}
-
-fn code(value: AuthReasonCode) -> u8 {
-    match value {
-        AuthReasonCode::Success => 0x00,
-        AuthReasonCode::ContinueAuthentication => 0x18,
-        AuthReasonCode::Reauthenticate => 0x19,
     }
 }
 
@@ -219,13 +222,13 @@ mod test {
         // Use user_properties to pad the size to exceed ~128 bytes to make the
         // remaining_length field in the packet be 2 bytes long.
         let auth_props = AuthProperties {
-            authentication_method: Some("Authentication Method".into()),
-            authentication_data: Some("Authentication Data".into()),
-            reason_string: None,
+            method: Some("Authentication Method".into()),
+            data: Some("Authentication Data".into()),
+            reason: None,
             user_properties: vec![(USER_PROP_KEY.into(), USER_PROP_VAL.into())],
         };
 
-        let auth_pkt = Auth::new(AuthReasonCode::ContinueAuthentication, Some(auth_props));
+        let auth_pkt = Auth::new(AuthReasonCode::Continue, Some(auth_props));
 
         let size_from_size = auth_pkt.size();
         let size_from_write = auth_pkt.write(&mut dummy_bytes).unwrap();